+++ vitest.config.mts
@@ -3,6 +3,6 @@ import { defineConfig } from 'vitest/config'
 export default defineConfig({
   test: {
     // ...
-    setupFiles: ['./test/setup-env.ts', './test/setup-github.ts', './test/setup-octokit.ts'],
+    setupFiles: ['./test/setupEnv.ts', './test/setupGithub.ts', './test/setupOctokit.ts'],
   },
 })
